hp.configure: Don't fail if cannot delete array
If no array is present, the delete command fails. In such, don't
consider the deletion error as a show-stopper since this could happen.

Note that if the delete fails for other reasons, the program will be
stopped by other errors that will ocurrs at creation time which are not
protected by a try/catch.