Makefile: New test target
The 'make test' command will spawn a kvm with the following setup:
- one nic 'user' with
 * PXE booting from the tests/tftpboot directory with pxelinux 5.10
 * NAT to the network
 * Port Forwarding from localhost:2222 to ssh
- one other nic
- one virtio 2GB virtual disk called kvm_storage.img
- 512MB of ram
- serial line printed on the console stdio

Aside this VM, the host will start the following services :
- rsyncd on port 1515 to serv the install/ path
- http server with CGI support using Python on port 9000

Configuration files are generated on the fly, process are deleted once
the run_kvm.sh is stopped. That very handy as no local configuration has
to be done to get the services.

Note that a temporary edeploy.conf is generated locally but symlinked in
/etc. Another will have to make this as a parameter from upload.py to
avoid touching /etc.

About the PXE booting, the vmlinuz & initrd are installed dynamically from the "pxe"
target.

The 2GB local storage is provisionned as qcow2.

So, a simple 'make test' is enough to start the kvm will all it's
required configuration & services. This could be used to simulate the
complete process while using a local setup. This could be used
everywhere with almost nothing.

Enjoy,

PS: the kvm is booting, the network configuration is ok, the curl works.
It still needs some specific config files from /config but that will be
done in another patch.