upload.py: Do not crash on update_cmdb
When running the code with a big amount of data >100 while the kernel
uses a slow console (like a ttySx), the priting of the computed ip list
trigger the execption and prevent upload.py from working.

This patch does simply comment out the printing which was only for
debugging purposes.