ipmi: Adding fake configuration & test for kvm
When running under kvm, let's create a fake ipmi device.

This could be used in addition with a network-class definition for ipmi
to insure the pairing is working properly.

That said, we're only expecting the following messages at boot time:

 Info: Added fake IPMI device
 Info: ipmi_setup_user: Setting user="root", password="pastouche" on channel 0
 Info: ipmi_setup_network: Setting network ip="10.0.4.5", netmask="255.255.255.0", gateway="10.0.4.2", vland_id="-1" on channel 0

It clearly show that an IP from pool 'ip-ipmi': '10.0.4.3-253' got
properly assigned to the fake IPMI device.