init: Reporting required kernel modules
When debugging some hosts, it's pretty hard to estimate why this or this
peripheral didn't worked as expected. In the current case this patch is
made for, we don't have network connectivity.

The first question that comes in mind is : does the network adapter have
a kernel driver in this kernel ?

To be more generic, we need to print (as we don't always have a console
on the targer host) the required kernel modules. We parse the pci
devices and report the associated kernel modules.

Note that this patch doesn't say if the expected kernel driver is
already builtin but at least, this does provides an helper.

Usually network card drivers & storage drivers are available as modules
so it should works nicely.