base.install: Switching to non-virtual kernel
Precise distribution was using virtual kernel leading to a shorten set
of drivers. This is a big deal as installing on real hardware requires
much more drivers than virtual environements.

As a real example, we were lacking of igb.ko on a supermicro server.

So let's switch to non-virtual kernels since it features much more
drivers for the real world. It have been checked that -virtual packaging
is a subset of regular kernels as no modules are missing.

This patch will make the image bigger but we need it.