init: Displaying dots while waiting eDeploy server
As long as curl is alive, let's plot a dot every second to indicate we
are waiting for something instead of being silent. This avoid the
feeling of having the system stuck.

If the edeploy server is here we do have a trace like :
 ################################################################
 Configuring system via server 10.0.2.2
 ################################################################
 Awaiting eDeploy server :....localhost.localdomain - - [09/Jul/2013 09:27:56] "POST //cgi-bin//upload.py HTTP/1.1" 200 -
 .done

Unless we do have:
 ################################################################
 Configuring system via server 10.0.2.2
 ################################################################
 Awaiting eDeploy server :.......................................

This is patch was done after Emilien's experience where he wasn't sure
about the reason why the boot was stuck at " Configuring system via
server x.x.x.x"