init: Massive rework of network detection
We had a real case where mii-tool was working on one interface but not
the other. Using the existing code, the second interface was not
considered for DHCP requests leading to a "no match".

This patch is reworking all this detection stuff.

mii-tool is deprecated, so let's use the /sys interface to detect
carrier. This is supposed to be filled by all network drivers.

This patch is also putting in parallel all DHCP requests to speed up the
detection time. Prior to this patch, the DHCP timeout were added for
each interface while now, it's a single wait for all interfaces.

As a result, network detection shall be faster with more reliability
regarding interface carrier detection.