IPMI: implement SOL support from the OS
With the help of IPMI and SOL (Serial OverLine) we can easily
access the console of the server if we need to troubleshoot it.
Thus this avoids the use of the JAVA console to debug the server.
Moreover we don't need any funcky iLO advanced licence for that.
Thanks to this patch the system automatically redirect its console
to a serial one.