hpacucli: Fixing ALL_SHOW_REGEXP
The ALL_SHOW_REGEXP was a little bit too much precise and prevented the
controller detection to occurs on some 420i.

We were expecting something like :
Smart Array P420 in Slot 2                (sn: PDKRH0ARH4F1IY)

While we got:
Smart Array P420i in Slot 0 (Embedded)    (sn: 5001438026807050)

In fact, regarding the location of the controller an "(Embedded)" string
could be added. This was preventing the \s* to match.