configure: Don't mkpart primary 0 xx
We do create partitions by using an optimal method to insure all IOs
will be aligned on the storage device.

For this, parted uses the optimal_io_size of the block device and align
the partition to that. As a result, if this size is 1MB big, the
checking of the optimisation fails and the /configure exit prematurely.

So it's better starting from 0% which include this computation.