upload.py: Print a specific message if run() fails
When run() call fails, the /configure is exiting prematurely but in
monk's silence.

That makes very hard to proof-read a /log file as the complete behavior of
/configure is weird while not having a proof about the reason.

This patch add a specific and very clear message that a particular
command failed and lead to an unexpected exit of configure.

That should make debugging easier.

A typical output will be :

	plop
	sh: 1: plop: not found
	Command 'plop' failed
	!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
	!!! Configure script exited prematurely !!!
	!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
	. /role
	/init: line 408: /role: No such file or directory