upload: Improving hw filename scheme
This patch change the way the hw filename is composed.

Generate a file name for a hardware using DMI information
(product name and version) then if the DMI serial number is
available we use it unless we lookup the first mac address.

As a result, we do have a filename like :
                    <dmi_product_name>-<dmi_product_version>-{dmi_serial_num|mac_address}