health-check: Reporting disk IOs performance
This patch add the ability to detect storage's performance by running
some fio patterns.

For a given test (4K read random or 1M read sequential), the test is
run of every disk in a standalone mode to detect the best of a device
without loading the system completly.

Then the test is run one more time on all detected's disks at the same
time to analyse how disks are performing when all the disks are
stressed. This could be useful to detect a controller not able to
substain the complete load