health: Adding upload feature
When using health mode with edeploy, it's useless being able to report
data to the server.

upload-health.py is a fork of upload.py but only keep the saving part of
it. No script is returned to the client.

A specific health/ directory got created to save the uploaded files.