upload-health: Saving results in timestamped dir
When uploading many profiles due to a big cluster run or in case of
multiple submission like in a smoke test, it's important to keep the
file on a timestamp basis to keep several versions of the same file.

As per the FAT directory of the USB bootable device, .hw files are
stored in a timestamp directory inside the heatlh/ dir.

A typical output is :
health/2013_09_18-10h24/