netdetect: Adding IP filtering
Before sending the synthesis to all nodes, we need to do some filtering.
Some servers are maybe in a different subnet and cannot participate in
the benchmark.

Let's count what subnetwork is getting the more servers from the initial
server list and then filter the server list to get one IP per server
than match this network.

Once the filtering is done, let's send it to the servers.

The next step is to self-eliminate servers that are no more part of this
list since they shall not participate in the benchmark.