netdetect: Adding MAX_CLIENTS support
When booting various kind of servers, it could be interesting telling
netdetect how many servers shall participate in the benchmark.

If the MAX_CLIENTS environement variable is setup (could be done via
kernel's cmdline also), servers will wait endlessly until this number of
clients is reached.

If MAX_CLIENTS is not defined, we keep the automatic detect & timeout approach.