netdetect: Don't lock if no server found
Prior this commit, the server didn't exited if no host got found.
To avoid this, when the client is considering that no other server can
be found, it sends a message telling the server to exit.

We also insure that in all case the hw.py provided is printed at the
end.

If more than 1 server is detected then the benchmarking stuff can start.