config: Freeing block device with dm_setup
When installing the same server multiple time, and if you are using some
LVM stuff you are facing this issue:

Selected disks, partition alignement & fstype (or LVM setup) is exactly
the same over and over.

This generate a problem as the kernel detects on the device the LVM
metadata that remained from the previous data. Detecting them, the
kernel is locking the device with some dm-* devices. This make then
impossible to complete some partitionning, formating or lvm actions.

As a solution, let's release all DMs before clearing the disks, let's
clean them with dd, reread the partition table to insure that we don't
keep ghost partitions, partition it again and _refree_ all DMs.

The second refree is important as if the LVM is not the first partition
the dd didn't had any effect if you keep the same data alignement.
Parted is so creating the same partitions at the same location and when
registering the partition, the dm-* device appear and relock the device.
The 2nd dm_setup remove_all does free the device in such case.

This is a very weird effect of reproductible installation on the same
hardware. Please consider this for every configure you create.