common: libmlx4 backport for Ubuntu precise
While using the verbs under Ubuntu precise, I had the following
behavior :

root@hpbig2:~# ibv_devices
libibverbs: Warning: no userspace device-specific driver found for /sys/class/infiniband_verbs/uverbs0
    device                 node GUID
    ------              ----------------
root@hpbig2:~#

As per https://bugs.launchpad.net/ubuntu/+source/libmlx4/+bug/1037107, and at the day of today,
the only solution we have is to install a package from quantal.

This is clearly weird to install a package this way but it makes verbs working as :

root@hpbig2:~# ibv_devices
    device                 node GUID
    ------              ----------------
    mlx4_0              0002c90300ea6820
root@hpbig2:~#

So this patch is taking the binary package and install it to the dest if we run a Ubuntu Precise.