functions: Using rsync --delete to copy tree
When developping some target, it's pretty common to run the make target
many many times. But the rsync wasn't using delete meaning that all new
files created by the target are not deleted between passes.

As a result, you can have a target that fails because of a previous run
and the rsync was not cleaning this part.

Using --delete make it works like a charm now.