infiniband: Compiling dkms during initial install
Prior approach wasn't the good one. In fact, if we need to install the
-generic headers package to insure a proper compilation.

This allow to generate the dkms at installation time. A few quirks were
required :
- linking the running kernel to the target one
- keeping a single kernel to avoid compilation against 3.8
- adding more links in arch/ subdir to make it happy (maybe related to
  the building env ? debian vs ubuntu ?)
- Adding noninteractive mode to apt to avoid errors at install time

With this patch the installation worked perfectly, it's now time to test
it by rebooting a physical server with an OS build with it.