init: Adding another root detection for kexec
The existing code that try to grab the root device from grub
configuration considers that the root device is UUID based.

But it does exist many other way to define a root device like :
- Labels
- Path

In such case, the root detection fails and prevents kexec from working.

This patch uses a 2nd detection if the first one fails and try to match
as much as possible the root device dependingless of the syntax. I
didn't push it as the default detection as I don't know it this could
work on other configs.