init: Uploading log file if init fails
While deploying servers with edeploy, it could be sometimes difficult to
understand why a server isn't installing properly and reboots.

To ease this debugging and while keeping DEBUG=0 to keep the same
decision path, we can now enable UPLOAD_LOG=1 option at the command
line or during the makefile when using bootable USB images.

When enabled, UPLOAD_LOG= will push the log file to edeploy server if
anything fails during the execution of init.

To specify where the log file shall be uploaded, use LOGDIR= directive in
edeploy.conf. The log file will be name with the vendor, product and
serial number of the host.