edeploy: Preemptive exit if no exclude file found
If no exclude file got found, rsyncing would destroy the host as the
rsync with delete mode will try to empty many local dirs as remote
server doesn't have them mounted.

This commit is also removing the fake exclude list during verify. This
is not a good idea to estimate that an hardcoded list of directories
will be ignored while it's not the case. This give a false perception of
what will be the real upgrade. If we find the exclude let's use it
directly but nothing else.

This commit is made for reinforcing the upgrade procedure to avoid
smashing systems while upgrading.