upgrade-from: Creating metadata structures
When creating an upgrade path, let's also create the default structures:
- /var/lib/debootstrap/* directories linked to this upgrade
- an exclude file filled with all the usual files and directories not to
  update

Without any exclude file or by trying to rsync empty dirs (not mounted)
or files created at boot time, we could have a very bad upgrade
scenario like killing /dev/ or grub.

This patch add a default series of files and directories to take
care-of. This is not an absolute list but the smallest one that should
match all linux distro. It have to be tuned by users but at least it
gives a good start.