Makefile: Allow {debian|ubuntu} repository override
When building images, it could be much more faster to use a local
repository instead of using the default one (in the US for debian).

This is done by using REPOSITORY= variable at make time like :

    make REPOSITORY=ftp://ftp.free.fr/pub/Distributions_Linux/debian
or
    export REPOSITORY=ftp://ftp.free.fr/pub/Distributions_Linux/debian
    make

Be careful that if you change the DIST= variable, REPOSITORY have to be
adjusted accordingly.

Enjoy,