upgrade-from: Make a real rsync when creating version
When creating a new version and especially while developping it, it's
important to restart from a clean base repository instead of the former
one.

This avoid having corrupted repository by having manually modified files
in the final repository in between two runs.

This doesn't affect performance on real use case but insure that
development trees are always clean