upgrade-from: /etc/ shall be managed file by file
In fact, excluding /etc was a mistake as adding a new service will add
files in it. On the other hand, we shall not touched files that are
specific to the host.

Let's try a first set of files that should not be touched to let the new
comers being added.