upgrade: Handling only added files
Upgrading is a little bit more complex than only rsyncing the tree and
excluding some directories.

Let's take an example where between V1.0 and V1.1 you add apache2.
To avoid losing logs, /var was excluded _but_ apache2 a new directory
and new files in /var/log/apache2.

This lead to a complex situation where allowing rsync on /var/log could
damage log files and preventing rsync lead to a failing apache2 at
startup time.

The proposed solution is to split the /var management in two phases :
- first we prevent rsync on /var/lib/edeploy, /var/cache & /var/log
- then, a second pass where we only allow new files to be added while
  not removing files

In this setup, log files are protected and new files are allowed.

To make this possible a new file named 'add_only' is used to list
directories that shall be considered as being able to add new files
while not considering the existing files. At the end, only new files are
added.

Using this patch I was able to add apache2 in an update and make it
work.