*.install: Updating to use packages functions
As part of story #3579, this commit change all existing roles to use new
functions defined in packages.

It makes the code more generic and easier to manage across distro.
That said, some quirks are still in but I let them for now to avoid
breaking too many stuff in a single shot.

Hope this patch is safe enough.