use a stamp for every target and .PHONY
This way make know the result is already up to date and won't
try to construct the target.

"make" uses the date of the stamp file to known when the last
build of the target has been done.

With this changes, make will check this date before building the
target and so avoid the execution of the same target two time in
a row.

A .PHONY target is one that is not really the name of a file; rather it
is just a name for a recipe to be executed when you make an explicit
request. There are two reasons to use a phony target: to avoid a
conflict with a file of the same name, and to improve performance, see:
http://www.gnu.org/savannah-checkouts/gnu/make/manual/html_node/Phony-Targets.html