common: Adding generic mount_fs call
As we need to setup the same fs twice at least (one for base, one for
the others), let's make a generic mount_fs function to avoid code
duplication.

More than that a bug was part of the mounting iso semantic. Since we
umount everything in between jobs, the ISO needs to be remounted as the
same time as the virtual fs. Let's manage this case inside the
mount_fs() function.