packages: Adding add_satellite_repository()
When SATELLITE_USERNAME and SATELLITE_PASSWORD variables are set, when
running a Redhat OS, the Redhat Network Satellite is configured to get
updates and additional packages.

i.e : puppet package requires rubygems which is only available on the
optional repository.

This patch configure the RHN regarding the user & password given.

Note it requires the "System Group Administrators and Satellite
Administrators" to give the proper rights to SATELLITE_USERNAME.