Add version and role to edeploy CLI
When one is dealing with several guests it's hard to keep track on
the version/role of each guest. Looking at the conf file is counterintuitive.
Since we source the conf file at the beginning of edeploy script version and role
are know. This PR simply display them via `edeploy version` and `edeploy role`.