init: Adding more logs inside the upload log
When uploading the log file, it's pretty handy to get much more about
the context of the host we deploy. Like if the configure is failing but
the host is unreachable, getting the /configure or the dmesg could help
to understand the problem.

So we add both /configure and dmesg when uploading the log file.