base: Removing system from satellite after building
When building an RHEL system, we have to use satellite.

But we don't want to
- saturate the number of deploied system
- let the system configured with an enovance account

So for each built target, we need to setup the satellite and remove it
at the end of the build. This way, we don't consume too much deployed
system on the satellite and we don't let configured systems.

On the other hand, we are not keeping a trace of the system inside the
satellite web interface making more difficult to track if updates are
available for a given role.

This is based on a python script found on a mailing list and adapted to
our needs (remove_system_on_satellite). It adds a perl-Frontier
dependency but it saves time regarding porting it in python.

Note also that calls to unregister_satellite() &
add_satellite_repository() are called from cleanup() and common_setup().
This implies that signal trap shall be enabled as soon a possible to
avoid some virtual FS (/dev/ etc..) mounted if this calls fails.

Last point about this commit, a small rework regarding the cleanup()
code to avoid the case statement. That made lots of code duplication and
very tough to read. Instead of that, we use a variable named
(STRIPPED_TARGET) to tell cleanup() not to try remove packages in it.