common: Do not mount /dev/ in bind mode
When /dev is let unmounted, like under kill -9 or an untrapped event, we
are in a situation where the /dev is leaved as mounted in the chroot.

When another job is coming to clear it, the /dev of the host is damaged.

This patch simply copy the content of the /dev/ in chroot's /dev.
As we have only special files, this is not an issue and no more risk to
kill the /dev's host.