clean: Removing useless roles
This commit is about removing unsupported roles.

Many roles got created on test purpose or before creating the
openstack-full role. To get a faster jenkins build and to keep the
smallest amount of supported roles, let's remove them.

If we need them back at some point, it will be easier to restart from
sample.* or from the openstack-full.

This will reduce the amount of work when touching roles & functions.