common: Don't Virtualize build dependencies
When building virtualized images with VIRTUALIZED= variable, it's
useless to generate virtualized images of all dependencies.

Prior this patch, if 'base' role was rebuilt, a virtual machine with the
same properties as the final role was created. This was taking a long
while but also make the build process longer.

This patch aims at detecting when the current target we build is the
last one (the one given on the command line).

MAKEFILE_TARGET variable own what the user requested to build while
CURRENT_TARGET variable represents the current target we build.

Once all the build dependencies are built, MAKEFILE_TARGET equals
CURRENT_TARGET and so we can consider creating the virtualized image.