detect.py: Adding net device businfo
While trying to keep a constant interface naming, it is useful to have
the bus path which represent the PCI path to the device.

This PCI path is supposed to be fixed across the same HW product and so
could be used to detect a physical port on a system.

A typical output will be:
    ('network', 'eth0', 'businfo', 'pci@0000:00:19.0'),

Note that on VMs this parameter will not be populated.