init: Adding multiple kernels support with kexec
It's pretty common to have several kernel version installed on the
deployed OS. When rebooting with grub this isn't a big issue since grub
will ask a boot time the kernel you want.

On the other hand, when booting with kexec, the current code was
considering that a single kernel was available. In case of multiple
kernels present on the /boot the code was wrong and prevented the system
to boot. A kernel panic occured.

To fix this situation, the current code is doing the following:
- If a single kernel exists, the behavior is unchanged, kexec will boot
  it
- If 2 or more kernels exists :
  -- by default the first one in the list will be booted
  -- user can enforce the kernel version to use by defining KEXEC_KERNEL= variable

The following example show how to boot a particular version of the
kernel :
LABEL eDeploy
	KERNEL vmlinuz
	INITRD initrd SERV=10.0.2.2 ONSUCCESS=kexec KEXEC_KERNEL=3.2.0-4