init: Reusing console= with kexec if none already defined
When testing or deploying systems, we usually setup some console=
parameters on the bootstrap commandline. This is very useful to see
kernel booting or logs on a serial line or Serial-Over-Lan.

If you choose onsuccess=kexec, thoses settings where lost meaning that
at the time the kexec could executed the serial line was not more used.

This patch checks that if some console= parameters are used for the
bootstrap _AND_ if no console parameters where already defined on the
BOOT_ARG then, we reuse those set to boot the bootstrap to spawn the
kexec kernel.

As a result, the kexec'd system print its console log of the same as the
bootstrap. Very handy.