init: Adding stop_log
The kill $tpid in save_log was an issue since it killed the logging once
save_log was called.

We had a case where the kexec stuff was not logged properly since
save_log killed tpid.

Frederic told me that tpid have to be killed to get kexec working
properly.

This patch add a stop_log() function that kills tpid. It is called very
late to get as much log as possible.