Revert "build/common: disable apt-get and yum outside of build scripts"
This reverts commit 3b6563958d7a26109dbc43ff1aa98c19a9f0513f.

closes: #57

Since 3b65639 , apt-get and yum are renamed to apt-get.moved and
yum.moved at the end of the builds.

When a role depends on another one, say base, the common_setup() function
must be called to restore these links.

Sadly, this common_setup() is not called in most of the roles. This
breaks the build.