health & pxe: Installing DEPS into ramfs
The makefile was defining some DEPS for both health & pxe but the
respective .install files were putting a different set of files in the
ramfs.

As a result, the health image was lacking of megacli.py.

So let's pass the DEPS variable to the install script to insure the same
files are used. That removes a useless and boring duplication.