create-image: ensure the root fs get resized
On Debian and Ubuntu, we use cloud-initramfs-growroot to get the
root file system resized.
1. The partition get resized in the initrd wth growroot
2. We call resize2fs in the /etc/rc.local to extend the ext3 fs

growroot is called directly from the initrd. It cannot do the
partition <> UUID look at this stage of the boot process. Since
we already know the name of the root partition (vda1), we hard
code it here.