health: Adding SESSION support
When performing benchmark, it's pretty common to change settings between
several phases to later analyze them.

The current behavior was about uploading everything by date turning into
a difficult task to remember what host was running which setup at what
time.

To ease this, a SESSION parameter have been added. When a result is
uploaded to the usb stick or on the edeploy server, a directory is
created with the SESSION name. All the files are then located into this
dir.