cardiff: Adding log-level selection
This patch allow users filtering the kind of messages they want to see.
The log-level is not a level you expect, but the kind of message you
want by adding them.

A typical usage is to get only the ERROR messages :
        cardiff.py -p 'sample/*.hw_'   -l ERROR

A typical usage is to get only the ERROR and WARNING messages :
        cardiff.py -p 'sample/*.hw_'   -l WARNING,ERROR

By default, we do print all the levels like a -l INFO,WARNING,ERROR
would had set.