init: Spliting probe_network_devices
This function was both setting link up and do the DHCP stuff.

To be able making an installation on a system that doesn't have any DHCP
server, we'll have to set the interface up but not expect any DHCP from
it.

This patch split the existing function in 2:
- enable_network_links
- probe_network_devices

enable_network_links:
This function does setup the link up and wait a little bit to let time
carriers to be set up. If the "fatal" argument is passed, then not
detecting any link will be considered as a fatal error

probe_network_device:
This funciton does the DHCP probing.