img: Adding CMDLINE support
On some host, we need to put some arguments on the command line to make
the system booting. If we need this hack at install time, it clearly
means we need it at rebooting time.

If user adds the CMDLINE= at build time, those arguments will be reused
on the kexec or on grub to insure the rebooted system to feature the
same tricks as the installer used.