detect: Reporting logical|megaraid|hpa disk's count
As per issue #84, we need to report the disk's count.

Server A have 1 disk : 500GB
Server B have 2 disks of the same type of server A

Server A & B are really the same the but the disks. If you do a rule
that match a single disk, both servers will match.

This patch add three new values :

-> disk, logical, count
count the number of logical (/dev/{s|h|v}d*) disks

disk, megaraid, count
count the number of phyiscal disks on the megaraid controllers

disk, hpa, count
count the number of phyiscal disks on the HP controllers