bootstrap: Earlier grabbing of edeploy-roles
As per issue #77, we can have issues at reporting which profile fails as
the information was arriving too late.

This patch aims at printing as soon a possible inside the configure
script the name of the profile that matched. Even if the configure is
totally broken, this information will be printed to the configure file.

This way, the edeploy client will be in position to reincrement the
role's counter even if the configure was totally broken.