health: Better detection of the USB mode
As per issue #82, it's important that AHC insure that if he boots on USB
and no AHCEXPORT partition is found, it shall stop.

In destructive mode, the usb key would be considered as a writable block
device ... which is wrong but in non-destructive the results would not
be saved on the USB ....

So let's stop AHC if no ahcexport partition found when booting on a USB
device.

So I'm just exporting a USBBOOT parameter on the commandline when
building the image with img.install. IF the USBBOOT parameter exists, it
means that we booted on an USB device. Not 100% perfect, but it does the
job.