detect: Adding Raid10 support for megacli
When megacli reports a Raid10 device, the NumberOfDrives is not
populated.

So not having a NumberOfDrives shall not trigger an exception but we
shall try to get it if available.

At the same time, we are saving NumberOfDrivesPerSpan to get more
details about this Raid10.

This commit solve the issue found on some real RAID10 setup.