init: Adding extlinux quirk for HP-DL120-G7 & Grub1
Some HP servers like HP DL120-G7 have generates invalid opcode with
Grub1 (like in Redhat 6.5).

This sounds to be related to this issue : http://lists.gnu.org/archive/html/grub-devel/2011-12/msg00017.html

In 2014, even with an up to date server (bios/firmwares) we still
encounter this issue when loading big kernels in grub1 if we got a
LOCALBOOT before.

The workaround here is not to use the grub1 anymore but the extlinux
instead. This setup is much more stable in our configuration.

This patch is so installing extlinux even if grub got already setup &
configured. That could sound weird but it makes the code much more easy
to read and also reusing many commonalities in both like kernel
detection & distro names.