init: Adding EOL at the end of the log
Adding an easy to remember string (End Of Log) at the end of the logs
will ease users to find the last string of the log.

Up to now, it was a boring scrolling from the bottom, through dmesg &
configure to get the last message.

Searching for EOL is now faster.