init/health: Use scl python27 for Redhat & Centos
Prior to this patch, it was impossible to run detect.py on a Redhat or
Centos derivative since the 6.5 release was using python 2.6.

Python 2.6 wasn't able to handle such kind of xml code :
   for elt in xml.findall(".//node[@id='firmware']"):

This patch is adding a python 2.7 backport on those version and add a
wrapper to run detect.py if such configuration is detected.

This patch makes Redhat & Centos being supported.