detect: Adding more information about the OS
It could be pretty useful to keep a trace on which kind of distribution
& kernel version the hw.py got generated.

This patch add some information about the os vendor & version but also
on the kernel version, booted architecture and command line.

This will give much more context if some hardware descriptions were
incomplete, wrong or missing.

A typical output would be:

 ('system', 'os', 'vendor', 'Mageia'),
 ('system', 'os', 'version', 'Mageia 4'),
 ('system', 'kernel', 'version', '3.12.13-desktop-2.mga4'),
 ('system', 'kernel', 'arch', 'x86_64')
 ('system', 'kernel', 'cmdline',
       'BOOT_IMAGE=linux root=UUID=44f8fb0d-ed65-4835-bc00-a1e5e81b2b73
       splash quiet resume=UUID=d8b3f967-8950-4b3b-8c6a-764d232429b3
       vga=788')