detect: Mac addresses shall be reported later
It's a common task to create spec files based on hw.py files. Up to now,
the serial (mac addr) was the first reported item.

This line is usually used to save the mac addr of a matched interface.
This saved value is used later then in the configure script.

It's so a common mistake to let the mac addr as the first line of the
set of rules defining one interface. This is not an enough
discriminative criteria.

Moving it to the last detected element will avoid people making wrong
spec files.