pkg_diff: Improving output
This patch format the output to ease the reading when comparing the version.
A fixed column approach is used, package name are left-aligned while version are right-aligned.

Note that if a package is new, a N/A is shown in the previous release.

Enjoy,

 ### Version changes ###
Package name                       Previous version             Newer version
------------------------- ------------------------- -------------------------
nova-api                                     2014.1                    2014.2
cinder-api                                      N/A                    2014.2