base: Fixing /var/run & /run issue
The issue is not at dbus level but at the filesystem one.
In fact, the /var/run shall point to /run. If the /var/run is a
directory it means that yum isn't patched and created /var/run while it
is supposed to be a symlink as the filesystem package is trying to do.

This is still a crappy hack while waiting #bz1130939 to be solved on
RH7.