health-server: Adding affinity-hosts support
It's now possible to define a list of hosts that shall be used for a
benchmark. This is very useful when running ahc in distributed mode to
selects VMs from a particular hypervisor.

Inside a job, you can use the comma-separated 'affinity-hosts' option to
select the host you want to use during this test.

The affinity-hosts is made of host's UUIDs like :
    affinity-hosts : 44454c4c-4b00-1039-8050-b9c04f573032