health-server: Measure jittering
This patch implements the first measure of jittering, it does measure:
- the time a host takes to ack the start bench command
- the time the test actually ran

This will be useful to estimate if all the VMs started the benchmark at
the same time and if the benchmark did last as long as expected.

As we do measure everything from the server side, if a host is
under-scheduled, the time required to perform the bench will show some
distortion.

We now need some code to reject the benchmark if values are not
acceptable