init: Adding cloud-init support to get configuration
When booting the bootstrap or the benchmarking tool, it is possible
using cloud-init to pass some arguments

To get this feature working, just add a user-data that features an #EDEPLOYMAGIC tag followed
by a file that will be sourced as a bash script.

So just put some VARIABLE=<value> stuff in it.

Note that all variables specified in this user-data file will override
any value passed at boot time via the command line.

Note also that we do retry 10 times while waiting 5 secs cloud-init to
answer. I saw many cloud-init call being stuck for a long while and
failing. So this implementation make the assomption that we could reach
some cloud-init configuration by trying 10 times for 5 seconds while
waiting 2 sec in between. The maximum time we wait for a cloud-init
configuration before failing is 66 seconds.