health: Implementing distributed memory bench
This commit introduce an option into the yaml file to setup distributed
memory benchmarks.

The format of the yaml file is so changing a little as all jobs to run
shall be under the "jobs" node as shown in dahc/test.yaml

Some functions related to the cpu bench have been renamed to insure a
proper naming as sysbench provides both cpu & memory test.

This work got used on a prototype with fancy results.

Note that I do increment the protocol_version since I added some new
variable in the messages.