cardiff: Plotting expected value when possible
If we do know the expected value for a given bench (like a job
duration), we can plot it directly to see how close to this value we
are.

This is providing more information to the reader to estimate how good is
the result.

The "variance" name got also renamed to "deviance".