openstack/common: Disable package even when enable is not the first keyword
Currently, services are disabled only if enable or disable is the first
keywork in the systemctl command. For example :

  systemctl enable myservice.service

Unfortunately, enable or disable isn't always the first word. For example,
Puppet enable its service with this command :

  systemctl --no-reload enable puppet.service

Hence the current script does not work anymore. This patch makes the
`install_packages_disabled` works no matter where is the enable keyword
located.