health: Adding MCE status reporting
When doing benchmark, this could lever some MCE exceptions. If so, we
shall catch them and report them.

By default the resulting file will be like :
    ('system', 'mce', 'generated', 'False')]

If an MCE got triggered during the test, we'll have it switch to True
like :
    ('system', 'mce', 'generated', 'True')]

Obviously, we shall not have _any_ mce triggered during a test.